Definition of the hierarchical event handling The HierarchicalInteractor defined the base behavior of a hierarchical interactor.
The StateMachinXML:
The modes of an interactor have to be set in the transition of an interactor:
The action <action ID="1100" /> set the mode not selected
The action <action ID="1101" /> set the mode selected
The action <action ID="1102" /> set the mode sub selected

The example behavior "HirachicalInteractor" in StateMachineXML shows the principles of a hierarchical state machine. It is important for the state sub selected that there is a transition with the id 1020 and the action <action ID="1101" /> or <action ID="1102" />. The action set the sub selected mode in selected or not selected mode. The event with the id = 1020 will send from the hierarchical interactor to itself when the sub element are deselected.