- FIX: Visual Studio 2005 issue:
STL Vector class changed slightly:
The application worked fine with VC7.1 since apparently the STL that shipped with VC7.1 did *something* to return a reference. The STL that ships with VC8.0 has conditional compilation and a serious of checks to trap this problem. One of those tests is if the internally managed pointer for "first" is null.