FIX (#1185): Tests for SliceBasedSegmentation crash sometimes
In this case something was broken in mitkSetRegionTool (calculation of an initial memory offset inside some segmentation). 
This code is quite ugly, was copied from an earlier version. I added a fix. 
Some time in the far future, when all tools will be converted to use ITK, this code will be obsolete...