FIX (#1405): Normals rendering (hedgehog) for Geometry2D objects not cut properly
Corrected hedgehog cutting by disabling calculation of clipped surface part. 
This had led to a complete normals data set (normals for both the clipped
and the non-clipped part).

In fixing this bug I discovered vtkActor's BackfaceProperties, which easily can 
and now is be used to color front and back side of a plane in different colors.