FIX (#1210): crash during loading a certain image
This issue was traced to a numerical issue of 
ITK-Code/Utilities/vxl/core/vnl/algo/vnl_svd.txx

Different compiler flags would not fix this behavior.
To fix this, a generated C translation of a Fortran algorithm
in ITK would have to be investigated further.

For now, I inserted two tests in Geometry3D to detect the
error, along with error handling code in ChiliPlugin.