FIX (#4991):  Possiblity to use multiple managers
Some MITK classes easily use the mitk::RenMan::GetInstance() singleton.
This limits the RenMan usage to the global instance and is prone to errors.

If an application uses some render windows with varying data in differing parts
of the application, we want to have the possibility to manage the rendering
process independent for each window.

Thus, the rendering manager should always be instantiated by means of:
mitk::BaseRenderer::GetInstance(vtkRenderWindow*)->GetRenderingManager()