FIX (#5177): allowing ScalarBar to become visible again
If you zoom in that far that the ScalarBar gets bigger than the RenderWindow it is set invisible. Right now, it cannot become visible again if you zoom out again. With this fix the ScalarBar will become visible again as soon as it's size is less than 70% of the size of the RenderWindow