Corrected spacing units (cm and mm)
In the Philips3D Dicom files, the spacing values are usually measured in centimeters.

The Tags
<0x0018,0x6024> (for x)
<0x0018,0x6026> (for y)
<0x3001,0x1002> (for z)
describe the measuring unit for the different spacings. If there is the value "3", this means, it is centimeters.
Since all Philips3D files, I encounterd are measured in centimeters, I do not know, if there is any other possible value for these tags.

The ITK Image Spacing is in milimteres, so I have to multiply the spacing values from the dicom header by 10 (if it is centimeters). I added an if-query for that.

Also I changed the index of the itk image to 0,0,0. Before, it was filled with the wrong values.