fixed a bug and added new features in NavigationDataRecorder
bugfix: stream is now flushed before it is deleted, so everything is written to disk before closing the stream. Also deleted the stream because it is a C pointer (this wasn't done before).

new features: now there is a method to get the number of logged/recorded frames, you can now also set a limit of this number.