Regiongrower is now working also in coronal window
This bug is strongly related to bug 7079. The cause is that the geometry of the
2D slice obtained by the ExtractImageFilter is wrong.

Because of that the current (clicked) worldposition is transformed into index
coordinates incorrectly. The RegionGrower then checks whether the index coordinates
lie within the current 2D slice. This is not so for the coronal window.
As a result of that the regiongrower calculated the wrong last screenposition.
This causes the strange cursor behaviour.

We now changed again (as in Bug #7079) the initialization of the PlaneGeometry
for the extracted slice. If the Geometry3D based on which the PlaneGeometry should
be created is an image geometry then the 0,5 offset is substracted from the new
origin.

Our change is:

If we are considering the x,y-plane then 0,5 should only be substracted from
the x and y coordinate and NOT from the z component.

If we are considering the y,z-plane then 0,5 should only be substracted from
the y and z coordinate and NOT from the x component. Analoguos for the x,z-plane.

We tested all tools in all windows and even the slice interpolation. It all
seems to work fine.