PlanePositionManager takes now a Geometry2D instead of a RestorePlanePositionOperation
Change the PlanePositionManager so that AddNewPosition() now takes a Geometry2D
instead of an RestorePlanePositionOperation.

Removed the function RestorePlanePosition from the SliceNavigationController.
Now the ExecuteOperation function must be called in order to apply a RestorePlanePositionOperation.

Changed the name of the contour markers. Now they are called just Position and
they count from 1.