Work in progress
Removed the "Remember Contour Positions" checkbox from the segmentation view.

Enhanced the SurfaceInterpolationController so that the 3D contours can be
displayed in the 3D window.

Added a new option for the interpolation: "Disable" so that no interpolation
is performed.

Provided the possibility for the 3D interpolation to hide the contours and the
marker nodes and of course to show them again.

The interpolation process is now executed in a pipeline.