Implemented Save/Load Tool Option
Saving is performed by writing a .ntf file with the tool information on harddisk.
If saving successfull a signal is emitted with the id of the tool and the filepath of the savefile.

If loading of an .ntf file is successfull the index and the a DataNode for the tool are emitted.