The level/window fields accept and display high precision values - fixes #8
With this fix you can enter values with more then two decimal digit
precision into the level/window text fields. Moreover, if the values
are between the allowed range then they will not be changed because
of double->float conversion.

The fields accept values in both exponential and fixed format,
but they will display the values in the format that has been previously
set by the SetExponentialFormat/SetPrecision function.

In a usual application it is determined by the "Exponential format" and
"Precision" preferences of the StdMultiWidget plugin.