Optimized paintbrush tool
Now the slice in which should be painted is stored in a data node as soon as
the user moves the mouse cursor into a renderwindow.

When the user presses the left mouse button and moves the mouse the contour
is filled into the 2D slice.

When the user releases the left mouse button the slice is written back into
the image volume