Moved separate example views to a central bundle
To clean up the CMake variables it has been decided to collect related views in one bundle. We have done this for the diverse examples for developing with MITK. This did include porting a few as-of-yet unported bundles to CTK.
We also changed the doxygen documentation to reflect this new structure, by changing the related pages dependencies.
The new central bundle has no central view, but a central doxygen page, linking to the constituent bundles.