corrected exports
Removed exports from templated classes and included .txx files at the end of the header. Because linking fails when no lib file exists, the only non templated class was given an export macro. The resulting lib is not necessary, but as far as I see it the only way for succesful linking in visual studio.