Improved upper and lower TH calculation
Now the THs are calculated depending on the RG direction:
Generally the max. range of the intensity values is estimated and therefrom
the windowSize for the RG which is 20 percent of the max. range.
For upwards RG direction:
lower TH = seedValueMean - 0.15*windowSize and upper TH = seedValueMean + 0.85*windowSize
For downwards RG direction:
lower TH = seedValueMean - 0.85*windowSize and upper TH = seedValueMean + 0.15*windowSize