Improved UI, fixed threshold calculation and improved volume visualization
UI: Removed checkbox for specifying threshold. Now if a seedpoint is placed the autmatically calculated threshold is displayed in the spinboxed. But the user can always specifiy an arbitrary upper or lower threshold.

Fixed some minor issues in the automatic threshold calculation

Afte performing regiongrowing the volume visualisation was not shown correctly