Fixed crash if multiple segmentations are loaded without reference image
The reason for this bug was that even if no reference image was loaded a initialized
reference data node was set as reference data for the tool manager. However the
data object this node hold was null. If no data is available the data node should not
be initialized too. This is now fixed.