Fixed crash of segmentation plugin when a segmentation for a segmentation should be created
Reason for this bug was an unused variable in the SegmentationView. To create
a segmentation of a segmentation the original segmentation has to be treated a
grayvalue image. Therefor the "binary" property has to be unchecked.
A segmentation is always handled as workingdata in the toolmanager. when the
binary flag is changed it is handled as reference data somehow this lead to a crash.