Imitate VTK viewport behavior in OpenGL mapper rendering
When VTK gets a viewport via vtkRenderer::SetViewport()
it scales the scene into this viewport. This commit makes
MITK's OpenGL based mappers bahave in the same way.

As commented in mitk::VtkPropRenderer, this is not an ideal
solution but a first fix. Plus, there are quite some tests
to save us from regressions.