Reword a "strange special case" to "ExpectDistanceOne" and change its implementation
This commit adds a flag m_ExpectDistanceOne to DICOMTagBasedSorter, which handles
a special case during "strict sorting". Documentation was enhanced to describe
this.

The default for "strict sorting" has been changed, the new default is
"no strict sorting". XMLs have been changed, so that we use strict sorting
for Instance Number and Slice Location now.