Test framework to enforce atomic state transitions
The tester can be connected to ITK events or Qt signals of
the test object or any other object whose state should be
consistent with the test object. E.g. if you are testing
the QmitkStdMultiWidget, you might want to connect the tester
to the ITK events coming from its renderers, slice navigation
controllers, the focus manager or the display geometries.

When a signal is received, the tester checks if any state transition
has happened actually, and if new state is consistent. The
test also refuses the state transitions that happens after
receiving a signal. No signal should be sent out until the
state of the object is 'final'. Intermediate states should not be
visible.

The commit contains a unit test for QmitkStdMultiWidget.