Fixed wrong loading of labelset image
The exterior label is always serialized when a mitkLabelSetImage is saved.
If the image is loaded again first of all a new layer is created which contains a exterior label as well.
Then the labels are added starting with the exterior.
Since this label already exists it will get the value +1 (current behavior of AddLabel).
This lead to a wrong ordering of the labels and thus a wrong visualization.