New preferences for crosshair gap size and corner annotations. CRASH WHEN CLOSING.
Added color and text for corner annotations of the 4 render windows to the StdMultiWidget preferences. Also crosshair gap size. There is still a crash when closing the workbench in the VtkLayerController. The controller holds a list of all renderers in a vector. I changed the annotation renderer to a smartpointer which is now automatically deleted. This is however, not propagated to the mitkVtkLayerController.