Fix Imagegeometry-Bug for Telemed-Ultrasoundimage
When the depth of the Ultrasoundimage was changed, the geometry of the
image remained the same. This was caused by the fact that the geometry of
the raw image takes a few frames to adapt. So if you update the geometry
of the ultrasoundimage to early it will be updatet with the old geometry.
Now the geometry of the ultrasoundimage is only updated when the geometry
of the raw image has changed.