Fix Imagegeometrybug for TelemedUltrasoundimage
When depth of an ultrasoundimage was changed the geometry remained the
same. This was caused by the fact, that the geometry of the raw image
takes a few frames to adapt to the new depth and the geometry of the
ultraosundimage was updated to early. Now the geometry of the
ultrasoundimage updates after the geometry of the raw image has changed
and therefore is correct.