Fixes freezebutton issues in MarkerPlacement
The freezebutton in the step placement planning didn't work correctly. It
did not unfreeze the image when clicked while image was already frozen. Also
the image did not unfreeze after adding a new zone. This was due to not
checking if image was already frozen in the method OnFreeze()
In the step MarkerIntervention when clicked on the freezebutton MITK crashed.
This was because when loading the state machine the wrong us::ModuleRegistry
was loaded.
Signed-off-by: Dominique Cheray <d.cheray@dkfz-heidelberg.de>