Integrates EchoTrack into actual master
org.mitk.gui.qt.usnavigation now depends on IGTIO because of the
NavigationDataSetWriterXML and the NavigationDataSetWriterCSV in this
module. Also the using of those two in mitkUSPointMarkInteractor had to be
adapted to the chanings made in the two writers.
mitkUSZonesInteractor and mitkUSPointMarkInteractor had to be adapted to
the changes made in their superclass.