Completes Calibration of EchoTrack with PLUS
You can now calibrate EchoTrack using PLUS and save the Calibration for
later use. PLUS provides the complete ImageToTrackerTransform, so no
calculating in MITK is needed. Just recieve the Transform and save it,
There is still work to do, to use all this with an UltrasoundMachine that
is connected using a framegrabber. For now it just works with the Telemed
device.