Dept and probe controls for USVideoDevice
Expands the USVideoDeviceCustomControls by a possibility for setting the
current scanning depth that is used and by a possibility for setting an
identifier for the used probe. This is helpfull when using EchoTrack with
an USVideoDevice and you need to save or load the correct calibration. You
can now set the scanning depth and the probe used with the ultrasound
machine manually in MITK and it is used for saving or loading the correct
calibration of EchoTrack.