Calibrate spacing of USVideoDevice
When calibrating EchoTrack you can now choose to determine the spacing of
the image of the USVideoDevice first. This spacing is then applied to the
USImagw. When changing the depth of the probe the spacing will no longer
be applied to the image because new depth means new spacing.
And the VirtualTracker that was always created when USNavigationPlugin was
used is now removed.