Fixes Workbenchcrash when closing USCalibration
Cause of crash was a vector in QmitkToolTrackingStatusWidget which was
sometimes allocated with new sometimes not. But it is always deleted.
Therefore you may try to delete something youd do not need to delete. Now
QmitkToolTrackingStatuisWidget has a flag that becomes true, when new was
used and the vector is only deleted if new was used.