Fixes minor fault in SpacingCalibration
When NacvigationData was invalid and Freezebutton in Spacing Calibration
was clicked the image would not freeze but you could still add points.
This was becaused the method from the Freezebutton in the PointBased
Calibration was used. There you need trackingdata to calibrate and
therefore it makes no sense to freeze the image without valid
trackingdata. But for the Spacing Calibration you don't need trackingdata so
this needed to be changed.
Also Freezebutton in Spacing Calibration is now disabled when all four
point are collected. There was some typo and the wrong button was
disabled.