Adds methods to USVideoDevice to get probes
Two new methods for USVideoDevice to get all probes or to get currently used probe. This is the first step in the changing of the USVideoDevice to simplify its use. Long term goal is to add and remove probes to the USVideoDevice and store them to. And each probe will get its depths and associated spacings. But at the moment there are just these two methods. Still much work is left.