extends saving and loading of USVideoDevice
Now the probes (including their scanning depths and spacings) of an
USVideoDevice are saved and reloaded. Next step is to change the GUI of
the USVideoDevice so the loaded probes and depths are shown and can be
selected. After this step the creation of a new USVideoDevice has to be
changed so a new probe can be created together with its scanning depths.
And calibration needs to be changed so the spacing will be saved to the
device.