Flag to suppress processing command line args
The flag is called 'applicationArgs.processByMITK'. If the flag is not
defined or it evaluates to 'true', the arguments are processed by MITK,
just like until now. It is possible, however, that custom applications
set this flag to 'false', so that they can introduce new switches or
options. (They could remove the dependency on the org.mitk.gui.qt.ext
module, but then they could not reuse other classes from it.)

To suppress processing the command line arguments, you need to put
this line into the main cpp file of your application:

  myApp.setProperty("applicationArgs.processByMITK", false);

where 'myApp' is typically an mitk::BaseApplication instance.

Signed-off-by: Miklos Espak <m.espak@ucl.ac.uk>