Finished Modification of USVideoDevice
Now an USVideoDevice can be edited. This means you can add or remove
probes. Change the Manufacturer, the Model or the Comment of the Device.
And remove or add scanning depths for the probes of the device.
Next step is to modifiy the calibration so the calibratet spacing for a
device is saved. And to modify the imaging so the calibrated spacing is
applid to the image.