Changed download URLs in dashboard scripts
For now we're using GitHub to download the files for the most common case of building the master branch. When another git branch is specified, we must download via Phabricator because we do not push other branches to GitHub. The reason for this differentiation is that downloading a repository file from Phabricator is an expensive operation, that generates temporary files in the Phabricator file application.