Fixes USVideoDevice not working on Linux
Reason for this problem was a deadlock in the Update of the ServiceProperties of the USDevice. The GUI
for the USVideoDeviceCustomControls updatet some Properties before the Device was completely activated.
Had to change the place where some signals and slots for the handling of probes and depths of the
USVideoDevice are connected. Now it works fine for Linus. Should be no problem on Windows becaus there
it was never a problem.