Starting on implementing functionality in the evaluation view
Finished the Gui. It is now nicely looking and by default filled with the
values needed to evaluate a standard .csv file from MITK.
Started to implement the functionality. For now a new
 mitk::NavigationDataCSVSequentialPlayer will be filled with the options
from the gui when constructed.
Next step is to use the new options when reading the .csv file.