Create axial origin at world bottom not the top
The axial plane is rotated so that the displayed image is with eyes up.
Hence, it is not possible to put the axial origin in the world bottom
left back corner, as for the other orientations. The origin must be in
the front so that the image is rendered with eyes up.

However, it is still possible to leave the origin on the bottom plane,
rather then on the top. This commit corrects this.