Workaround bug T19825: IGT's TrackingVolumeGenerator expects to load surfaces but will crash in certain configurations.
Following scenario:
* a project registers another MIME type and its reader for STL files
* the specific type and reader get a higher service ranking, so will be preferred by IOUtil

mitk::TrackingVolumeGenerator calls IOUtil::Load(filename) and expects that the result
is a Surface. Even when the filename is an STL file (as is for all current devices), Load()
will use the first MIME type / reader that can handle STL files. When this is not the
MITK core service that produces mitk::Surface, TrackingVolumeGenerator will crash subsequently.

This commit provides a workaround for the current situation. The proper solution would
be to call IOUtil::LoadSurface() _after_ fixing the behavior of that method.

Signed-off-by: Daniel Maleike <code@maleike.de>