Block selection model signals when selection not set through interaction
The data manager view listens to changes in the tree view selection model so that
when the user clicks on a data node, its 'selected' property is updated accordingly.

However, this slot that updates the 'selected' properties is also called when the
selection is set by another workbench part. In this case the data manager updates
the properties incorrectly, because it supposes that the data manager view is the
active workbench part, and it takes the current selection from the active part,
instead of the data manager view itself.

This fix blocks the signals from the selection model when it is updated from code,
not by user interaction.

Signed-off-by: Miklos Espak <m.espak@ucl.ac.uk>