Remove MitkSimulation
MitkSimulation isn't compatible with the OpenGL2 Backend of VTK. As
MitkSimulation is based on a rather old version of SOFA and never left
the experimental stage, we remove it completely for now.

The integration would have to be written from scratch anyway for a recent
version of SOFA. If we want to have simulation capabilities based on SOFA
again we have to make sure SOFA isn't using any fixed-function pipeline
OpenGL code anymore.