Remove ConstructTrackingDevice and replace by GetTrackingDevice.
Only the concrete implementation of a tracking device should decide,
if a tracking device needs to be freshly constructed each time (compare NDI Aurora) or
if the tracking device can/needs to be stored in a member and returned (compare Polhemus).

This needs adaption of ALL previous tracking devices, e.g. external code!