Add classes to create a patient table model / inspector based on the semantic relations.
A subclass of the node selection dialog is used to provide a caseID for
the patient table inspector. The patient table inspector is a data
storage inspector but it uses a special semantic relations model that
retieves its data not from the data storage but from a semantic
relations storage.
The patient table inspector is added as a service reference via the
semantic relations ui module activator.