Refactor data manager
Summary:
Moved functions of the context menu to own classes and created a context menu class that initializes these actions.
Some actions contain a globally accessible function (namespace) to provide the functionality without initializing the corresponding action.
Preferences for 'GlobalReinit' have been moved to the 'GeneralPreferencePage'.

Had to remove the 'QmitkNodeTableViewKeyFilter' and the 'QmitkDataManagerHotkeysPrefPage', since the actions where not accessible anymore.
If needed the required actions can be made globally accessible.

Also fixes T22435 and T25254.

Some topics remain unresolved:
- T25176 and the related T25177
- T18414

We still have to discuss which actions are needed in the context menu and which not.

Test Plan: manual testing

Reviewers: kislinsk

Subscribers: kalali

Maniphest Tasks: T25056, T25254, T22435

Differential Revision: https://phabricator.mitk.org/D156