Add two custom actions
Move the action functionality from the patient table inspector to these
actions.
Extend the semantic relations file to allow setting the information type
and sending a notificitation for semantic relation observers to update
their storage view.
The semantic relations view now implements the
'IRenderWindowPartListener' to set the controlled renderer if the render
window part is activated.