Fix opacity changes of segmentations
The opacity of segmentations is reset when opening the segmentation
views. This is done with different values for different conditions.
However, whatever the original intention was, the condition does not
change as expected and hence the opacity is always reset to 1.

When this fix breaks a feature, consider introducing a second opacity
property for this purpose and multiply the opacities in the mapper.
Thus the opacity explicitly set by the user wouldn't get overriden.