Allow usage of module-external interactions in tools
If a module is specified as optional parameter in the Tool constructor,
the state machine is loaded from that module. It is also assumed that
there's a event config file in the same module with the same name,
suffixed by Config.xml. For example, Paint.xml have to be accompanied
by PaintConfig.xml and the type parameter of the constructor is "Paint".