Compute statistics with the help of the image statistics ui module.
The ui module contains a QmitkImageStatisticsCalculationJob that can be
used to start a thread that computes the lesion volume.
The approach is taken from the image statistics view and adapted to the
semantic relations needs.
This leads to statistics not computed multiple times and allows to
continue working while the statistics are being computed.