When changing a property's value, to NOT call Modified() on the PropertyLIST
With the current logic, the model will rebuild itself whenever the
propertylist changes. This will in consequence collapse a tree view
that displays the model.

That behavior is not nice but can be tolerated when something else in
the application triggers a modification of the propertylist. When the
user modifies a child-item via the 'Properties' view then this
collapsing is very annoying, though.

Signed-off-by: Daniel Maleike <daniel.maleike@stryker.com>