Check filename for invalid characters before attempting to save.
Without such a check writers might (and some do) create empty files
that are not even named as expected (e.g. "a" instead of "a:b.obj").

The opening std::ofstream for such a filename does not seem to set
any fail-bits (all checked for in the affected writer class).

The alternative to a check here would be a check in each possible writer
class. This seems worse.

Signed-off-by: Daniel Maleike <daniel.maleike@stryker.com>