AbstractFileReader stores relevant IO meta information (BREAKING!)
This is a interface breaking commit for all classes
derived from AbstractFileReader. Derived classes have to
implement now "std::vector<BaseData::Pointer> DoRead()".
Normaly the only thing to do, is to rename the Read()
implementation of the derived class to DoRead().

The change was needed to ensure that for all reader classes
the meta information is stored. This is done now in the
implementation of AbstractFileReader::Read() after DoRead()
is called.

Signed-off-by: Ralf Floca <r.floca@dkfz-heidelberg.de>