Fixed T26979
Save guard that ensures that the bounds are always valid for a plane geometry
(bounds[1] and bounds[3] must not be zero).
If this is the complete fix or only curing the symptoms,
will be discussed in the task. Depending on the discussion
results this may not be over.

The problem occures when the current world geometry, that
should be rendered by the mapper, has an FOV that in combination with
a loaded image generates an index (right or bottom vector) that is
smaller then 1 and will be rounded to 0, wich violates the bound checks
of plane geometry.
More details in Task T26979

Signed-off-by: Ralf Floca <r.floca@dkfz-heidelberg.de>