Changed ExtractSliceFilter::m_WorldGeometry from raw to smart pointer
I think it is unnecessary unsafe to have the member as unmanaged raw pointer,
as we use it in the filter and have no posibility to check if it is still valid.
Therefore cleaned up and changed it to smart pointer even though it is not
the scope of this task.

Signed-off-by: Ralf Floca <r.floca@dkfz-heidelberg.de>