Fixed T27259
It is still the problem explained in https://phabricator.mitk.org/T24766#131411.

I have thought about different possibilities to solve it; e.g.
# Ensure that we have no collapsed final time steps.
# Support infinite time bounds for the last time bounds.
# Introduce the support of the corner case of a collapsed final time step in ArbitratyTimeGeometry. In this case we accept time points as valid and within the last time step if the equal the max time bound

Option 1 I have dismissed, because it would mean to introduce a rule to deduce the duration of the last time step for most dynamic DICOM images as it is not stored. This is not trivial and can introduce new pit falls (and we somehow alter the dicom data which is also not good). 
Option 2 I have dismissed, as for dicom it is just a special version of option 1 with all strings attached.  And DICOM images are what causes the pain right now.

So I have opted for option 3 to fix T27259. It seems the minimal invasive but still acceptable clean solution for the problem.

Signed-off-by: Ralf Floca <r.floca@dkfz-heidelberg.de>